requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.2
Pillow==10.1.0
reportlab==4.0.8
openpyxl==3.1.2
pytz==2023.3
